test(client): move to prismaPromise
