+++ packages/client/src/__tests__/integration/happy/prismaPromise/.gitignore
@@ -1,0 +1,1 @@
+!dev.db
