+++ packages/client/src/__tests__/integration/happy/prismaPromise/schema.prisma
@@ -1,0 +1,66 @@
+datasource db {
+  provider = "sqlite"
+  url      = "file:dev.db"
+}
+
+generator client {
+  provider      = "prisma-client-js"
+  binaryTargets = ["native"]
+}
+
+model User {
+  id         String   @id @default(cuid())
+  email      String   @unique
+  name       String?
+  age        Int?
+  posts      Post[]
+  likes      Like[]
+  property   Property? @relation(fields: [propertyId], references: [id])
+  propertyId String?
+  Banking    Banking?
+}
+
+model Property {
+  id      String @id @default(cuid())
+  house   House  @relation(fields: [houseId], references: [id])
+  users   User[]
+  houseId String
+}
+
+model House {
+  id         String     @id @default(cuid())
+  like       Like       @relation(fields: [likeId], references: [id])
+  properties Property[]
+  likeId     String
+}
+
+model Post {
+  id        String   @id @default(cuid())
+  createdAt DateTime @default(now())
+  updatedAt DateTime @updatedAt
+  published Boolean
+  title     String
+  content   String?
+  authorId  String?
+  author    User?    @relation(fields: [authorId], references: [id])
+  Like      Like[]
+}
+
+model Like {
+  id     String @id @default(cuid())
+  userId String
+  user   User   @relation(fields: [userId], references: [id])
+  postId String
+  post   Post   @relation(fields: [postId], references: [id])
+
+  House House[]
+  @@unique([userId, postId])
+}
+
+model Banking {
+  id     String @id @default(cuid())
+  userId String
+  user   User?  @relation(fields: [userId], references: [id])
+  iban   String
+  bic    String
+}
